Separate package version and Git resource functions from create release
