+++ pkg/cmd/release/create/create_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/OctopusDeploy/cli/pkg/cmd/release/create"
 	cmdRoot "github.com/OctopusDeploy/cli/pkg/cmd/root"
 	"github.com/OctopusDeploy/cli/pkg/executor"
+	"github.com/OctopusDeploy/cli/pkg/packages"
 	"github.com/OctopusDeploy/cli/pkg/surveyext"
 	"github.com/OctopusDeploy/cli/test/fixtures"
 	"github.com/OctopusDeploy/cli/test/testutil"
@@ -21,7 +22,7 @@ import (
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/credentials"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/deployments"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/feeds"
-	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/packages"
+	octopusPackages "github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/packages"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/projects"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/releases"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/resources"
@@ -213,8 +214,8 @@ func TestReleaseCreate_AskQuestions_RegularProject(t *testing.T) {
 			}})
 
 			// then it will look for versions
-			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-				Items: []*packages.PackageVersion{{PackageID: "pterm", Version: "0.12.51"}},
+			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+				Items: []*octopusPackages.PackageVersion{{PackageID: "pterm", Version: "0.12.51"}},
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{
@@ -260,7 +261,7 @@ func TestReleaseCreate_AskQuestions_RegularProject(t *testing.T) {
 
 			var fireProject2 = *fireProject // clone the struct value
 			fireProject2.VersioningStrategy = &projects.VersioningStrategy{
-				DonorPackage: &packages.DeploymentActionPackage{
+				DonorPackage: &octopusPackages.DeploymentActionPackage{
 					DeploymentAction: "Verify",
 					PackageReference: "nuget-on-verify",
 				},
@@ -318,11 +319,11 @@ func TestReleaseCreate_AskQuestions_RegularProject(t *testing.T) {
 			}})
 
 			// then it will look for versions
-			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-				Items: []*packages.PackageVersion{{PackageID: "pterm", Version: "0.12.51"}}, // extra package to prove it's not just picking the first one
+			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+				Items: []*octopusPackages.PackageVersion{{PackageID: "pterm", Version: "0.12.51"}}, // extra package to prove it's not just picking the first one
 			})
-			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-				Items: []*packages.PackageVersion{{PackageID: "NuGet.CommandLine", Version: "6.2.1"}}, // the proper package
+			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+				Items: []*octopusPackages.PackageVersion{{PackageID: "NuGet.CommandLine", Version: "6.2.1"}}, // the proper package
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{
@@ -371,7 +372,7 @@ func TestReleaseCreate_AskQuestions_RegularProject(t *testing.T) {
 
 			var fireProject2 = *fireProject // clone the struct value
 			fireProject2.VersioningStrategy = &projects.VersioningStrategy{
-				DonorPackage: &packages.DeploymentActionPackage{
+				DonorPackage: &octopusPackages.DeploymentActionPackage{
 					DeploymentAction: "Verify",
 					PackageReference: "nuget-on-verify",
 				},
@@ -414,8 +415,8 @@ func TestReleaseCreate_AskQuestions_RegularProject(t *testing.T) {
 			}})
 
 			// then it will look for versions
-			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-				Items: []*packages.PackageVersion{{PackageID: "NuGet.CommandLine", Version: "6.2.1"}}, // the proper package
+			api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+				Items: []*octopusPackages.PackageVersion{{PackageID: "NuGet.CommandLine", Version: "6.2.1"}}, // the proper package
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{
@@ -704,7 +705,7 @@ func TestReleaseCreate_AskQuestions_VersionControlledProject(t *testing.T) {
 
 func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
-	baseline := []*create.StepPackageVersion{
+	baseline := []*packages.StepPackageVersion{
 		{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 		{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "6.1.2"},
 		{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
@@ -716,8 +717,8 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 	}{
 		// this is the happy path where the CLI presents the list of server-selected packages and they just go 'yep'
 		{"no-op test", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("y")
@@ -725,12 +726,12 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
 			assert.Equal(t, baseline, versions)
-			assert.Equal(t, make([]*create.PackageVersionOverride, 0), overrides)
+			assert.Equal(t, make([]*packages.PackageVersionOverride, 0), overrides)
 		}},
 
 		{"override package based on package ID", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("pterm:2.5")
@@ -739,19 +740,19 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "6.1.2"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageID: "pterm", Version: "2.5"},
 			}, overrides)
 		}},
 
 		{"override package based on step name", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("Install:2.5")
@@ -760,19 +761,19 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "2.5"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{ActionName: "Install", Version: "2.5"},
 			}, overrides)
 		}},
 
 		{"override package based on package reference", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("Install:pterm:2.5")
@@ -781,12 +782,12 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "6.1.2"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "2.5"},
 			}, overrides)
 		}},
@@ -794,20 +795,20 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 		{"entering the loop with --package-version picked up from the command line", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
 			defaultPackageVersion := "2.5"
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, defaultPackageVersion, make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, defaultPackageVersion, make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("y")
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "2.5"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{Version: "2.5"}, // TODO the "regenerate command line flags" code is going to re-interpret this as "--package *:2.5" rather than the input which was "--package-version 2.5". Does that matter?
 			}, overrides)
 		}},
@@ -815,20 +816,20 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 		{"entering the loop with --package picked up from the command line", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
 			cmdlinePackages := []string{"Install:pterm:2.5", "NuGet.CommandLine:7.1"}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", cmdlinePackages, qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", cmdlinePackages, qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("y")
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "7.1"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "2.5"},
 				{PackageID: "NuGet.CommandLine", Version: "7.1"},
 			}, overrides)
@@ -838,20 +839,20 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 			defaultPackageVersion := "9.9"
 			cmdlinePackages := []string{"Install:pterm:2.5", "NuGet.CommandLine:7.1"}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, defaultPackageVersion, cmdlinePackages, qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, defaultPackageVersion, cmdlinePackages, qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("y")
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "7.1"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "9.9"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{Version: "9.9"}, // TODO the "regenerate command line flags" code is going to re-interpret this as "--package *:9.9" rather than the input which was "--package-version 9.9". Does that matter?
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "2.5"},
 				{PackageID: "NuGet.CommandLine", Version: "7.1"},
@@ -859,8 +860,8 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 		}},
 
 		{"blank answer retries the question", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			validationErr := qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("")
@@ -875,12 +876,12 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
 			assert.Equal(t, baseline, versions)
-			assert.Equal(t, make([]*create.PackageVersionOverride, 0), overrides)
+			assert.Equal(t, make([]*packages.PackageVersionOverride, 0), overrides)
 		}},
 
 		{"can't specify garbage; question loop retries", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion})
@@ -905,19 +906,19 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "2.5"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{Version: "2.5"},
 			}, overrides)
 		}},
 
 		{"can't specify packages or steps that aren't there due to validator; question loop retries", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion})
@@ -933,19 +934,19 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "2.5"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{Version: "2.5"},
 			}, overrides)
 		}},
 
 		{"question loop doesn't retry if it gets a hard error", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWithError(errors.New("hard fail"))
@@ -957,8 +958,8 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 		}},
 
 		{"multiple overrides with undo", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("NuGet.CommandLine:7.1")
@@ -973,20 +974,20 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "7.1"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"}, // this would have been hit by pterm:35 but we undid it
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageID: "NuGet.CommandLine", Version: "7.1"},
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "2.5"},
 			}, overrides)
 		}},
 
 		{"multiple overrides with reset", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baseline, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			_ = qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion}).AnswerWith("NuGet.CommandLine:7.1")
@@ -1001,26 +1002,26 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "2.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "6.1.2"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"}, // this would have been hit by pterm:35 but we undid it
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "2.5"},
 			}, overrides)
 		}},
 
 		// this is the happy path where the CLI presents the list of server-selected packages and they just go 'yep'
 		{"if we enter the loop with any unresolved packages, force version selection for them before entering the main loop", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			baselineSomeUnresolved := []*create.StepPackageVersion{
+			baselineSomeUnresolved := []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: ""},                         // unresolved
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: ""}, // unresolved
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baselineSomeUnresolved, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baselineSomeUnresolved, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: "Unable to find a version for \"pterm\". Specify a version:"})
@@ -1055,25 +1056,25 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "75"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "1.0.0"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "75"}, // fully qualify packagereference+actionname to be sure
 				{PackageReferenceName: "NuGet.CommandLine", ActionName: "Install", Version: "1.0.0"},
 			}, overrides)
 		}},
 
 		{"if we enter the loop with any unresolved packages, forced version selection doesn't accept bad input", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			baselineSomeUnresolved := []*create.StepPackageVersion{
+			baselineSomeUnresolved := []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: ""}, // unresolved
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baselineSomeUnresolved, "", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baselineSomeUnresolved, "", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: "Unable to find a version for \"pterm\". Specify a version:"})
@@ -1099,24 +1100,24 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "25.0"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "25.0"}, // fully qualify packagereference+actionname to be sure
 			}, overrides)
 		}},
 
 		{"if we enter the loop with any unresolved packages, pick up --package-version before assuming they're unresolved", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			baselineSomeUnresolved := []*create.StepPackageVersion{
+			baselineSomeUnresolved := []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: ""},                         // unresolved
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: ""}, // unresolved
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baselineSomeUnresolved, "12.7.5", make([]string, 0), qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baselineSomeUnresolved, "12.7.5", make([]string, 0), qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: packageOverrideQuestion})
@@ -1131,26 +1132,26 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "12.7.5"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "12.7.5"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "12.7.5"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{Version: "12.7.5"}, // the --package-version input produces this as the first 'override'
 				// and that's all we did there
 			}, overrides)
 		}},
 
 		{"if we enter the loop with any unresolved packages, pick up --package before assuming they're unresolved", func(t *testing.T, qa *testutil.AskMocker, stdout *bytes.Buffer) {
-			baselineSomeUnresolved := []*create.StepPackageVersion{
+			baselineSomeUnresolved := []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: ""},                         // unresolved
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: ""}, // unresolved
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}
 
-			receiver := testutil.GoBegin3(func() ([]*create.StepPackageVersion, []*create.PackageVersionOverride, error) {
-				return create.AskPackageOverrideLoop(baselineSomeUnresolved, "", []string{"Install:pterm:12.9.2"}, qa.AsAsker(), stdout)
+			receiver := testutil.GoBegin3(func() ([]*packages.StepPackageVersion, []*packages.PackageVersionOverride, error) {
+				return packages.AskPackageOverrideLoop(baselineSomeUnresolved, "", []string{"Install:pterm:12.9.2"}, qa.AsAsker(), stdout)
 			})
 
 			q := qa.ExpectQuestion(t, &survey.Input{Message: "Unable to find a version for \"NuGet.CommandLine\". Specify a version:"})
@@ -1175,12 +1176,12 @@ func TestReleaseCreate_AskQuestions_AskPackageOverrideLoop(t *testing.T) {
 
 			versions, overrides, err := testutil.ReceiveTriple(receiver)
 			assert.Nil(t, err)
-			assert.Equal(t, []*create.StepPackageVersion{
+			assert.Equal(t, []*packages.StepPackageVersion{
 				{ActionName: "Install", PackageID: "pterm", PackageReferenceName: "pterm", Version: "12.9.2"},
 				{ActionName: "Install", PackageID: "NuGet.CommandLine", PackageReferenceName: "NuGet.CommandLine", Version: "75"},
 				{ActionName: "Verify", PackageID: "pterm", PackageReferenceName: "pterm", Version: "0.12"},
 			}, versions)
-			assert.Equal(t, []*create.PackageVersionOverride{
+			assert.Equal(t, []*packages.PackageVersionOverride{
 				{PackageReferenceName: "pterm", ActionName: "Install", Version: "12.9.2"},         // input commandline switch produces this output
 				{PackageReferenceName: "NuGet.CommandLine", ActionName: "Install", Version: "75"}, // our first question produces this
 			}, overrides)
@@ -1857,10 +1858,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		// octopusApiClient.NewClient fetches the root resource but otherwise BuildPackageVersionBaseline does nothing
@@ -1869,7 +1870,7 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{}, packageVersions)
+		assert.Equal(t, []*packages.StepPackageVersion{}, packageVersions)
 	})
 
 	t.Run("builds list for single package/step", func(t *testing.T) {
@@ -1889,10 +1890,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -1908,15 +1909,15 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "pterm", Version: "0.12.51"},
 			},
 		})
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm",
 				ActionName:           "Install",
@@ -1944,10 +1945,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -1955,7 +1956,7 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{}, packageVersions)
+		assert.Equal(t, []*packages.StepPackageVersion{}, packageVersions)
 	})
 
 	t.Run("builds list containing only selectable package/step when a different step has a fixed package version", func(t *testing.T) {
@@ -1983,10 +1984,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -2002,15 +2003,15 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "pterm", Version: "0.12.51"},
 			},
 		})
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm",
 				ActionName:           "Install",
@@ -2058,10 +2059,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -2082,14 +2083,14 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "pterm", Version: "0.12.51"},
 			},
 		})
 		// even though two steps use pterm, they're the same so we don't need to ask the server twice
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "NuGet.CommandLine", Version: "6.1.2"},
 			},
 		})
@@ -2097,7 +2098,7 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm",
 				ActionName:           "Install",
@@ -2166,7 +2167,7 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 			{
 				Tag:          "^pre$",
 				VersionRange: "[5.0,6.0)",
-				ActionPackages: []packages.DeploymentActionPackage{
+				ActionPackages: []octopusPackages.DeploymentActionPackage{
 					{DeploymentAction: "Install", PackageReference: "pterm-on-NOSUCHSTEP"}, // this should be ignored as PackageReference doesn't match
 					{DeploymentAction: "Cleanup", PackageReference: "nuget-on-cleanup"},    // this should match
 				},
@@ -2175,16 +2176,16 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 			{
 				Tag:          "^$",
 				VersionRange: "[9.9]",
-				ActionPackages: []packages.DeploymentActionPackage{
+				ActionPackages: []octopusPackages.DeploymentActionPackage{
 					{DeploymentAction: "InstallXYZ", PackageReference: "pterm-on-install"}, // this should be ignored as DeploymentAction doesn't match
 				},
 			},
 		}
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -2205,27 +2206,27 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "pterm", Version: "0.12.51"},
 			},
 		})
 		// even though two steps use pterm, they're the same so we don't need to ask the server twice
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "NuGet.CommandLine", Version: "6.1.2"},
 			},
 		})
 		// second request asks for different filters due to channel rules
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&preReleaseTag=%5Epre%24&take=1&versionRange=%5B5.0%2C6.0%29").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/Feeds-1001/packages/versions?packageId=NuGet.CommandLine&preReleaseTag=%5Epre%24&take=1&versionRange=%5B5.0%2C6.0%29").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{
 				{PackageID: "NuGet.CommandLine", Version: "5.4.1-prerelease"},
 			},
 		})
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm",
 				ActionName:           "Install",
@@ -2267,10 +2268,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -2286,16 +2287,16 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{}, // empty!
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{}, // empty!
 		})
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{}, // empty!
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=NuGet.CommandLine&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{}, // empty!
 		})
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm",
 				ActionName:           "Install",
@@ -2325,10 +2326,10 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 
 		channel := fixtures.NewChannel(spaceID, "Channels-1", "Default", "Projects-1")
 
-		receiver := testutil.GoBegin2(func() ([]*create.StepPackageVersion, error) {
+		receiver := testutil.GoBegin2(func() ([]*packages.StepPackageVersion, error) {
 			defer api.Close()
 			octopus, _ := octopusApiClient.NewClient(testutil.NewMockHttpClientWithTransport(api), serverUrl, placeholderApiKey, "")
-			return create.BuildPackageVersionBaseline(octopus, processTemplate, channel)
+			return create.BuildPackageVersionBaselineForChannel(octopus, processTemplate, channel)
 		})
 
 		api.ExpectRequest(t, "GET", "/api/").RespondWith(rootResource)
@@ -2344,14 +2345,14 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 		}})
 
 		// now it will search for the package versions (the first one said resolvable:true so we ask the server, even though there's a variable template)
-		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm-%23%7BOctopus.Environment.Id%7D&take=1").RespondWith(&resources.Resources[*packages.PackageVersion]{
-			Items: []*packages.PackageVersion{}, // empty!
+		api.ExpectRequest(t, "GET", "/api/Spaces-1/feeds/feeds-builtin/packages/versions?packageId=pterm-%23%7BOctopus.Environment.Id%7D&take=1").RespondWith(&resources.Resources[*octopusPackages.PackageVersion]{
+			Items: []*octopusPackages.PackageVersion{}, // empty!
 		})
 		// the other two packages have IsResolvable:false so we don't even try to look for them.
 
 		packageVersions, err := testutil.ReceivePair(receiver)
 		assert.Nil(t, err)
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{
 				PackageID:            "pterm-#{Octopus.Environment.Name}",
 				ActionName:           "Install",
@@ -2378,19 +2379,19 @@ func TestReleaseCreate_BuildPackageVersionBaseline(t *testing.T) {
 func TestReleaseCreate_ToPackageOverrideString(t *testing.T) {
 	tests := []struct {
 		name   string
-		input  *create.PackageVersionOverride
+		input  *packages.PackageVersionOverride
 		expect string
 	}{
-		{name: "ver-only", input: &create.PackageVersionOverride{Version: "0.12"}, expect: "*:0.12"},
-		{name: "action-ver", input: &create.PackageVersionOverride{ActionName: "Install", Version: "0.12"}, expect: "Install:0.12"},
-		{name: "action-ver-2", input: &create.PackageVersionOverride{ActionName: "Verify", Version: "6.1.2-beta"}, expect: "Verify:6.1.2-beta"},
-		{name: "pkg-ver", input: &create.PackageVersionOverride{PackageID: "pterm", Version: "0.12"}, expect: "pterm:0.12"},
-		{name: "pkg-ver-2", input: &create.PackageVersionOverride{PackageID: "NuGet.CommandLine", Version: "6.1.2-beta"}, expect: "NuGet.CommandLine:6.1.2-beta"},
-		{name: "pkg-action-ver", input: &create.PackageVersionOverride{PackageID: "pterm", ActionName: "Install", Version: "0.12"}, expect: "pterm:0.12"}, // this isn't valid, but if it did happen it should pick packageID
-		{name: "pkg-ref-ver", input: &create.PackageVersionOverride{PackageReferenceName: "pterm-on-install", PackageID: "pterm", Version: "6.1.2"}, expect: "pterm:pterm-on-install:6.1.2"},
-		{name: "action-ref-ver", input: &create.PackageVersionOverride{PackageReferenceName: "pterm-on-install", ActionName: "Install", Version: "6.1.2"}, expect: "Install:pterm-on-install:6.1.2"},
-		{name: "star-ref-ver", input: &create.PackageVersionOverride{PackageReferenceName: "pterm-on-install", Version: "6.1.2"}, expect: "*:pterm-on-install:6.1.2"},
-		{name: "pkg-action-ref-ver", input: &create.PackageVersionOverride{PackageReferenceName: "pterm", PackageID: "pterm", ActionName: "Install", Version: "1.2.3"}, expect: "pterm:pterm:1.2.3"},
+		{name: "ver-only", input: &packages.PackageVersionOverride{Version: "0.12"}, expect: "*:0.12"},
+		{name: "action-ver", input: &packages.PackageVersionOverride{ActionName: "Install", Version: "0.12"}, expect: "Install:0.12"},
+		{name: "action-ver-2", input: &packages.PackageVersionOverride{ActionName: "Verify", Version: "6.1.2-beta"}, expect: "Verify:6.1.2-beta"},
+		{name: "pkg-ver", input: &packages.PackageVersionOverride{PackageID: "pterm", Version: "0.12"}, expect: "pterm:0.12"},
+		{name: "pkg-ver-2", input: &packages.PackageVersionOverride{PackageID: "NuGet.CommandLine", Version: "6.1.2-beta"}, expect: "NuGet.CommandLine:6.1.2-beta"},
+		{name: "pkg-action-ver", input: &packages.PackageVersionOverride{PackageID: "pterm", ActionName: "Install", Version: "0.12"}, expect: "pterm:0.12"}, // this isn't valid, but if it did happen it should pick packageID
+		{name: "pkg-ref-ver", input: &packages.PackageVersionOverride{PackageReferenceName: "pterm-on-install", PackageID: "pterm", Version: "6.1.2"}, expect: "pterm:pterm-on-install:6.1.2"},
+		{name: "action-ref-ver", input: &packages.PackageVersionOverride{PackageReferenceName: "pterm-on-install", ActionName: "Install", Version: "6.1.2"}, expect: "Install:pterm-on-install:6.1.2"},
+		{name: "star-ref-ver", input: &packages.PackageVersionOverride{PackageReferenceName: "pterm-on-install", Version: "6.1.2"}, expect: "*:pterm-on-install:6.1.2"},
+		{name: "pkg-action-ref-ver", input: &packages.PackageVersionOverride{PackageReferenceName: "pterm", PackageID: "pterm", ActionName: "Install", Version: "1.2.3"}, expect: "pterm:pterm:1.2.3"},
 	}
 
 	for _, test := range tests {
@@ -2404,25 +2405,25 @@ func TestReleaseCreate_ToPackageOverrideString(t *testing.T) {
 func TestReleaseCreate_ParsePackageOverrideString(t *testing.T) {
 	tests := []struct {
 		input     string
-		expect    *create.AmbiguousPackageVersionOverride
+		expect    *packages.AmbiguousPackageVersionOverride
 		expectErr error
 	}{
-		{input: ":5", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
-		{input: "::5", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
-		{input: "*:5", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
-		{input: "*:*:5", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
-		{input: ":*:5", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
-		{input: "NuGet:NuGet:0.1", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet", ActionNameOrPackageID: "NuGet", Version: "0.1"}},
-		{input: "NuGet:nuget-on-install:0.1", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "nuget-on-install", ActionNameOrPackageID: "NuGet", Version: "0.1"}},
-		{input: "Install:nuget-on-install:0.1", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "nuget-on-install", ActionNameOrPackageID: "Install", Version: "0.1"}},
-		{input: "pterm:9.7-pre-xyz", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
-		{input: "pterm:55", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "55"}},
-		{input: "pterm::55", expect: &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "55"}},
-		{input: ":Push Package:55", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "", Version: "55"}},
-		{input: "*:Push Package:55", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "", Version: "55"}},
+		{input: ":5", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
+		{input: "::5", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
+		{input: "*:5", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
+		{input: "*:*:5", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
+		{input: ":*:5", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "", Version: "5"}},
+		{input: "NuGet:NuGet:0.1", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet", ActionNameOrPackageID: "NuGet", Version: "0.1"}},
+		{input: "NuGet:nuget-on-install:0.1", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "nuget-on-install", ActionNameOrPackageID: "NuGet", Version: "0.1"}},
+		{input: "Install:nuget-on-install:0.1", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "nuget-on-install", ActionNameOrPackageID: "Install", Version: "0.1"}},
+		{input: "pterm:9.7-pre-xyz", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
+		{input: "pterm:55", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "55"}},
+		{input: "pterm::55", expect: &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "pterm", Version: "55"}},
+		{input: ":Push Package:55", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "", Version: "55"}},
+		{input: "*:Push Package:55", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "", Version: "55"}},
 
-		{input: "pterm/Push Package=9.7-pre-xyz", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
-		{input: "pterm=Push Package/9.7-pre-xyz", expect: &create.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
+		{input: "pterm/Push Package=9.7-pre-xyz", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
+		{input: "pterm=Push Package/9.7-pre-xyz", expect: &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "Push Package", ActionNameOrPackageID: "pterm", Version: "9.7-pre-xyz"}},
 
 		{input: "", expectErr: errors.New("empty package version specification")},
 
@@ -2438,7 +2439,7 @@ func TestReleaseCreate_ParsePackageOverrideString(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.input, func(t *testing.T) {
-			result, err := create.ParsePackageOverrideString(test.input)
+			result, err := packages.ParsePackageOverrideString(test.input)
 			assert.Equal(t, test.expectErr, err)
 			assert.Equal(t, test.expect, result)
 		})
@@ -2448,139 +2449,139 @@ func TestReleaseCreate_ParsePackageOverrideString(t *testing.T) {
 func TestReleaseCreate_ResolvePackageOverride(t *testing.T) {
 	// this is the packageId:5.0 syntax
 	t.Run("match on package ID", func(t *testing.T) { // this is probably the most common thing people will do
-		nugetPackage := &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
 	})
 
 	// this is the stepName:5.0 syntax
 	t.Run("match on step name", func(t *testing.T) { // this is probably the most common thing people will do
-		nugetPackage := &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "Install", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "Install", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "Install", Version: "5.0", PackageReferenceName: ""}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "Install", Version: "5.0", PackageReferenceName: ""}, r)
 	})
 
 	// this is the packageRef:*:5.0 syntax
 	t.Run("match on packageRef", func(t *testing.T) {
-		nugetPackage := &create.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet-A"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet-B"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
 	})
 
 	// this is the *:5.0 or :5.0 or just 5.0 syntax
 	t.Run("match on just version", func(t *testing.T) { // this is probably the most common thing people will do
-		nugetPackage := &create.AmbiguousPackageVersionOverride{Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
 	})
 
 	t.Run("match on just version doesn't even need any packages to look at", func(t *testing.T) { // this is probably the most common thing people will do
-		nugetPackage := &create.AmbiguousPackageVersionOverride{Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{Version: "5.0"}
 
-		steps := make([]*create.StepPackageVersion, 0) // baseline
+		steps := make([]*packages.StepPackageVersion, 0) // baseline
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: ""}, r)
 	})
 
 	t.Run("match on action+packageRef before packageID", func(t *testing.T) { // this is probably the most common thing people will do
-		nugetPackage := &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "Verify", PackageReferenceName: "NuGet", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "Verify", PackageReferenceName: "NuGet", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "Verify", Version: "5.0", PackageReferenceName: "NuGet"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "Verify", Version: "5.0", PackageReferenceName: "NuGet"}, r)
 	})
 
 	t.Run("match on packageID+packageRef picks the first one where they are the same", func(t *testing.T) {
-		nugetPackage := &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", PackageReferenceName: "NuGet", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", PackageReferenceName: "NuGet", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet"}, r)
 	})
 
 	t.Run("match on packageID+packageRef picks the correct one where they are different", func(t *testing.T) {
-		nugetPackage := &create.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", PackageReferenceName: "NuGet-B", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{ActionNameOrPackageID: "NuGet", PackageReferenceName: "NuGet-B", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet-A"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet-B"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "NuGet", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
 	})
 
 	t.Run("match on packageRef wins over match on ActionName", func(t *testing.T) {
 		// we shouldn't get in this situation, but just in case we do :shrug:
-		nugetPackage := &create.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", ActionNameOrPackageID: "Cheese", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", ActionNameOrPackageID: "Cheese", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet-A"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet-B"},
 			{PackageID: "OtherPackage", ActionName: "Cheese", Version: "0.1", PackageReferenceName: "Cheese"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
 	})
 
 	t.Run("match on packageRef wins over match on PackageID", func(t *testing.T) {
 		// we shouldn't get in this situation, but just in case we do :shrug:
-		nugetPackage := &create.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", ActionNameOrPackageID: "Cheese", Version: "5.0"}
+		nugetPackage := &packages.AmbiguousPackageVersionOverride{PackageReferenceName: "NuGet-B", ActionNameOrPackageID: "Cheese", Version: "5.0"}
 
-		steps := []*create.StepPackageVersion{ // baseline
+		steps := []*packages.StepPackageVersion{ // baseline
 			{PackageID: "NuGet", ActionName: "Install", Version: "0.1", PackageReferenceName: "NuGet-A"},
 			{PackageID: "NuGet", ActionName: "Verify", Version: "0.1", PackageReferenceName: "NuGet-B"},
 			{PackageID: "Cheese", ActionName: "OtherAction", Version: "0.1", PackageReferenceName: "OtherAction"},
 		}
 
-		r, err := create.ResolvePackageOverride(nugetPackage, steps)
+		r, err := packages.ResolvePackageOverride(nugetPackage, steps)
 		assert.Nil(t, err)
-		assert.Equal(t, &create.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
+		assert.Equal(t, &packages.PackageVersionOverride{PackageID: "", ActionName: "", Version: "5.0", PackageReferenceName: "NuGet-B"}, r)
 	})
 }
 
 func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
-	standardPackageSpec := []*create.StepPackageVersion{
+	standardPackageSpec := []*packages.StepPackageVersion{
 		{PackageID: "pterm", ActionName: "Install", Version: "0.12", PackageReferenceName: "pterm-on-install"},
 		{PackageID: "pterm", ActionName: "Push", Version: "0.12", PackageReferenceName: "pterm-on-push"},
 		{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "5.4", PackageReferenceName: "nuget-on-install"},
@@ -2588,11 +2589,11 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	}
 
 	t.Run("apply wildcard override", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{Version: "99"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "99", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "99", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "99", PackageReferenceName: "nuget-on-install"},
@@ -2601,11 +2602,11 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply one override based on package ID", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{PackageID: "pterm", Version: "99"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "99", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "99", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "5.4", PackageReferenceName: "nuget-on-install"},
@@ -2614,11 +2615,11 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply one override based on step name", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{ActionName: "Install", Version: "99"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "99", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "0.12", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "99", PackageReferenceName: "nuget-on-install"},
@@ -2627,11 +2628,11 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply one override based on both package and step name", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{PackageID: "pterm", ActionName: "Install", Version: "99"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "99", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "0.12", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "5.4", PackageReferenceName: "nuget-on-install"},
@@ -2640,13 +2641,13 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply multiple overrides", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{Version: "0.1"},
 			{PackageID: "pterm", Version: "2"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Push", Version: "6"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "2", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "2", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "0.1", PackageReferenceName: "nuget-on-install"},
@@ -2655,13 +2656,13 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply multiple overrides; order matters", func(t *testing.T) {
-		result := create.ApplyPackageOverrides(standardPackageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(standardPackageSpec, []*packages.PackageVersionOverride{
 			{PackageID: "pterm", Version: "2"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Push", Version: "6"},
 			{Version: "0.1"}, // overwrites everything
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", Version: "0.1", PackageReferenceName: "pterm-on-install"},
 			{PackageID: "pterm", ActionName: "Push", Version: "0.1", PackageReferenceName: "pterm-on-push"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", Version: "0.1", PackageReferenceName: "nuget-on-install"},
@@ -2670,18 +2671,18 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 	})
 
 	t.Run("apply single override targeting only package-ref", func(t *testing.T) {
-		packageSpec := []*create.StepPackageVersion{
+		packageSpec := []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", PackageReferenceName: "pterm-on-install", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Push", PackageReferenceName: "pterm", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Verify", PackageReferenceName: "pterm", Version: "0.12"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", PackageReferenceName: "NuGet.CommandLine", Version: "5.4"},
 		}
 
-		result := create.ApplyPackageOverrides(packageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(packageSpec, []*packages.PackageVersionOverride{
 			{PackageReferenceName: "pterm", Version: "2000"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", PackageReferenceName: "pterm-on-install", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Push", PackageReferenceName: "pterm", Version: "2000"},
 			{PackageID: "pterm", ActionName: "Verify", PackageReferenceName: "pterm", Version: "2000"},
@@ -2691,18 +2692,18 @@ func TestReleaseCreate_ApplyPackageOverride(t *testing.T) {
 
 	t.Run("target both of package-ref:action where package referencename matches another package too", func(t *testing.T) {
 		// real bug observed with manual testing
-		packageSpec := []*create.StepPackageVersion{
+		packageSpec := []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", PackageReferenceName: "pterm-on-install", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Push", PackageReferenceName: "pterm", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Verify", PackageReferenceName: "pterm", Version: "0.12"},
 			{PackageID: "NuGet.CommandLine", ActionName: "Install", PackageReferenceName: "NuGet.CommandLine", Version: "5.4"},
 		}
 
-		result := create.ApplyPackageOverrides(packageSpec, []*create.PackageVersionOverride{
+		result := packages.ApplyPackageOverrides(packageSpec, []*packages.PackageVersionOverride{
 			{PackageReferenceName: "pterm-on-install", ActionName: "Install", Version: "20000"},
 		})
 
-		assert.Equal(t, []*create.StepPackageVersion{
+		assert.Equal(t, []*packages.StepPackageVersion{
 			{PackageID: "pterm", ActionName: "Install", PackageReferenceName: "pterm-on-install", Version: "20000"}, // only this one should be overridden
 			{PackageID: "pterm", ActionName: "Push", PackageReferenceName: "pterm", Version: "0.12"},
 			{PackageID: "pterm", ActionName: "Verify", PackageReferenceName: "pterm", Version: "0.12"},
