--- pkg/cmd/release/create/string_utilities.go
@@ -1,43 +1,0 @@
-﻿package create
-
-import "golang.org/x/exp/slices"
-
-// splitString splits the input string into components based on delimiter characters.
-// we want to pick up empty entries here; so "::5" and ":pterm:5" should both return THREE components, rather than one or two
-// and we want to allow for multiple different delimeters.
-// neither the builtin golang strings.Split or strings.FieldsFunc support this. Logic borrowed from strings.FieldsFunc with heavy modifications
-func splitString(s string, delimiters []rune) []string {
-	// pass 1: collect spans; golang strings.FieldsFunc says it's much more efficient this way
-	type span struct {
-		start int
-		end   int
-	}
-	spans := make([]span, 0, 3)
-
-	// Find the field start and end indices.
-	start := 0 // we always start the first span at the beginning of the string
-	for idx, ch := range s {
-		if slices.Contains(delimiters, ch) {
-			if start >= 0 { // we found a delimiter and we are already in a span; end the span and start a new one
-				spans = append(spans, span{start, idx})
-				start = idx + 1
-			} else { // we found a delimiter and we are not in a span; start a new span
-				if start < 0 {
-					start = idx
-				}
-			}
-		}
-	}
-
-	// Last field might end at EOF.
-	if start >= 0 {
-		spans = append(spans, span{start, len(s)})
-	}
-
-	// pass 2: create strings from recorded field indices.
-	a := make([]string, len(spans))
-	for i, span := range spans {
-		a[i] = s[span.start:span.end]
-	}
-	return a
-}
